add resume preview feature

+++ src/app/containers/edit-profile/edit-profile.component.ts
@@ -412,15 +412,39 @@ export class EditProfileComponent implements OnInit {
     }
   }
   removeResumeFile() {
-    this.profileService.deleteDocumentLibraryByEmail(this.userInfo.user_name).subscribe((res: any) => {
-      if (res['code'] === 0) {
-        this.resumeFile = null;
-        this.userProfile.document_library = null;
-        this.alertService.successAlert('Resume delete successfully');
+    this.alertService.confirmAlert('Are you sure?', 'Do you really want to remove resume?')
+      .then((res: any) => {
+        if (res.value) {
+          this.profileService.deleteDocumentLibraryByEmail(this.userInfo.user_name).subscribe((res: any) => {
+            if (res['code'] === 0) {
+              this.resumeFile = null;
+              this.userProfile.document_library = null;
+              this.alertService.successAlert('Resume delete successfully');
+            } else {
+              this.alertService.errorAlert('Delete failed');
+            }
+          });
+        }
+      });    
+  }
+  previewResume() {
+    const fileType = this.resumeFile.name.split('.').pop();
+    const reader = new FileReader();
+    this.profileService.getResume(this.userProfile.resume_url).subscribe((res: any) => {
+      const blob = new Blob([res], {type: "octet/stream"}),
+            url = window.URL.createObjectURL(blob);
+      if (fileType && fileType.toString().toUpperCase() === 'PDF') {
+
       } else {
-        this.alertService.errorAlert('Delete failed');
-      }      
-    })
+        const downloadLink = document.createElement("a");
+        downloadLink.href = url;
+        downloadLink.download = this.resumeFile.name;
+
+        document.body.appendChild(downloadLink);
+        downloadLink.click();
+        document.body.removeChild(downloadLink);
+      }
+    });
   }
   imageCropped(image: string) {
     this.croppedImage = image;
